refactored the course name page to move the app methods into separte folders and use the express router method to connect them
