+++ middleware/authenticate.js
@@ -1,6 +1,6 @@
 function authenticate(req, res, next) {
     console.log('Authenticating...');
     next();
-};
+}
 
 module.exports = authenticate;
