+++ package.json
@@ -20,9 +20,11 @@
   "dependencies": {
     "@hapi/joi": "^17.1.1",
     "config": "^3.3.1",
+    "debug": "^4.1.1",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "helmet": "^3.22.0",
-    "morgan": "^1.10.0"
+    "morgan": "^1.10.0",
+    "pug": "^2.0.4"
   }
 }
