+++ routes/render-home.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', (req, res) => {
+    res.render('index', { title: 'My Express App', message: 'Hello' });
+});
+
+module.exports = router;
