+++ views/index.pug
@@ -1,0 +1,6 @@
+html
+    head
+        title= title
+    body
+    h1= message
+
